requirements.txt
Babel==2.2.0
bcrypt==2.0.0
blinker==1.4
cffi==1.5.2
coverage==4.0.3
cryptography==1.3
decorator==4.0.9
enum34==1.1.2
et-xmlfile==1.0.1
Flask==0.10.1
Flask-Babel==0.9
Flask-Bcrypt==0.7.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
httplib2==0.9.2
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
jdcal==1.2
Jinja2==2.8
MarkupSafe==0.23
oauth2client==2.0.1
openpyxl==2.3.4
pbr==1.8.1
peewee==2.8.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==16.0.0
python-openid==2.2.5
pytz==2016.3
rsa==3.4
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.12
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
Tempita==0.5.2
virtualenv==15.0.1
Werkzeug==0.11.4
Whoosh==2.7.3
WTForms==2.1
